fix: model migration
Removes model config from storage backend. This means that we use
the fallback storage pool for all requests. This fixes the issue
where state was imported before dqlite for k8s. We can't attempt
to use dqlite before or during state importing. We could invert it
to import dqlite before state, but we also run into this as well.

In the end, this is dead code, just remove it.