feat(firewall): wire up firewaller to use ports domain
Wire up the firewaller api on the controller facade to be backed by
the DQLite ports domain.

This included substituting the DQLite watcher for the Mongo watcher.

This required a quick patch to the DQLite watcher to emit machine names
instead of machine uuids, since this data is consumed outside the
apiserver, so should be a natural key.

We also needed to wire up OpenedPortRangesForMachine. This was a little
more complex.

The OpenedPortRangesForMachine spec requires us to return not just the
endpoint with their opened ports, but the subnet cidrs that endpoint is
opened to.

However, endpoint bindings have not yet been implemented in DQLite. So
instead, I have created a new method on the service,
GetMachineOpenedPortsAndSubnets, which for now just returns the opened
ports with nil SubnetCIDRs. Once we have endpoint bindings wired up in
DQLite, this stub method will need to be expanded.

This has been added to the risk register, and a ticket has been added
to the appropriate epic for this work.