ci: remove limits on commit body length in CI
Testing this with a really long body. I am doing this because sometimes the dependabot commits have a really long body and it causes issues with the CI.
The length of the body is now a warning not an error.

Also remove the limit on the foote since this is also arbitrary. Testing with a really really long footer.