feat(caasfirewaller): wire up caasfirewaller api
Wire up caasfirewaller api to use the new ports service, backed by
DQLite.

Two endpoints are effected. WatchOpenedPorts, and GetOpenedPorts.

Unfortuantely somehow the watcher I had previously implemented to be
ready to plug in place to the caasfirewaller was completely incorrect.
The watcher we desire is a strings watcher, emitting events when any
port changes containing the names of the applications changed.

This means a re-write of the ports domain application ports watcher was
required.

We can also fulfill a TODO left in the caasfirewaller applicationworker.
Unlike the firewaller worker, the caasfirewaller worker spins up an
applicationworker for each application in a model. So this worker can
skip change events which don't change this worker's specific
application.

Delete legacy mongo ports watcher