feat: adds Capture() to errors library
Based on usage and feedback we have decided to look at adding a method
similar to the old juju/errors.Trace to this library. This will be
similar in semantics and offer a means for capturing the call site where
the func was called from.

This commit has deliberately avoided calling this new method Trace so
that we can identify what is using old techniques v new techniques.