refactor(filtering): wire up ports domain to filtering
Refactor the (very old) client facade status filtering logic to filter
by ports acquired from DQLite backed storage.

This refactor required two things. First, all our closure functions
needed to be made into methods, so that they have access to services on
the client. Next, a context needed to be added to the params of our
predicate, to be woven through to our calls to services.