test: generate ssh keys for new models
In 4.0, the user's SSH keys are no longer automatically added to new
models. This is causing usages of juju ssh in Bash tests to fail. To
get around this, every time we create a model in a test, we generate an
SSH key pair for it, and provide this as an argument to juju ssh as
required. (excluding k8s clouds which don't support this)