test: delete 'unit' Bash test suite
The only purpose of this test suite was to test the
set-application-base command, which was removed in #17682. We already
have a test (run_deploy_specific_series in the deploy suite) that
ensures that charm bases can be specified on deploy. Hence, we can
remove the test suite (or replace it with a no-op).