chore: fix log level when setting machine cloud instance on bootstrap
Before this patch, when trying to set the machine cloud instance data
for the bootstrap machine, we would only log (on ERROR level) if we
couldn't retrieve the machine UUID. Of course this could be a real
error, but it is also the normal path when on k8s since we don't have
machines.
Just moving the log level to DEBUG to inform that this is not supported
in k8s. Similar to what we do when we cannot reload spaces when not
supported on those clouds.