refactor(firewaller): wire up ports domain directly
The firewaller worker is a model worker, meaning it runs directly on the
controller. This means we can grant the firewaller worker access to the
domains directly, instead of going via an API.

Wire up domainServices to the firewaller manifold, and attach the
portService and machineService to the firewall worker. Replace calls to
the API to calls these services for the openedPorts concerns.

Remove the unused API facades from the controller facade. This does NOT
result in compatibility issues because this is the controller facade, it
is not exposed to anyone besides the controller.

NOTE: The apiserver OpenedMachinePorts facade contained a lot of
redundancy. It would query application endpoints and spaces in order to
calculate the subnet CIDRs to expose particular endpoint port ranges to,
and return this data to the caller.

However, the firewaller worker would just drop this information, instead
using it's own watchers and api to get spaces/endpoint bindings
information, and calculate these subnet cidrs itself. This is required
because the firewaller need to watch for changes to spaces AS WELL as
changes to opened ports.

Since we are calling the domain service itself, we can remove this
redundancy. This resulted in a stub method on the ports service being
unused, so it's been removed.