feat: remove orphaned secret revisions
If secret revisions have been orphaned, it's important to clean
them up, otherwise we can end up with lots of revisions. This can
in turn create a lot of pressure on the database.

The script provided cleans up older revisions, and will attempt to
remove older ones. If there is an error or the revision is not found
it will move on to the next one.