refactor: use one goroutine
To prevent potential race conditions because the executing function
is called in another goroutine, we move to one. There is a lot of
change in here, we have to use to contexts to ensure that the
goroutine is correctly cleaned up.

I've detailed the strategy in the comments, but it should hold true
for the lifetime of the execution of the function.