feat: use storage registry from dependency engine
The storage registry is based on the providers, we already have a
concept for that. Use that, rather than an adhoc mechanism.

The idea is to unwind the all of the parameters added to the domain
services. This is the first one, the next will be the secrets and
key updater.

Once this is done, all dependencies for the domain must be sought
by the dependency engine.