refactor(caasfirewaller): wire up ports domain directly
The caasfirewaller worker is a model worker, meaning it runs directly on the
controller. This means we can grant the caasfirewaller worker access to the
domains directly, instead of going via an API.

Wire up domainServices to the caasfirewaller manifold, and attach the
portService and applicationService to the caasfirewall worker. Replace calls to
the API to calls these services for the openedPorts concerns.

Removed the unused API facades from the controller facade. This does NOT
result in compatibility issues because this is the controller facade, it
is not exposed to anyone besides the controller.