feat: attach registry to storage service
To remove the ad-hoc storage registry access, the registry is
attached to the service for now. This is essentially a hack to allow
us to thread the correct underlying registry to the storage backend.
The storage backend is a bit of a mess, because it takes a state
and creates all its dependencies ad-hoc style. Instead it would be
preferable to pass them in. Attempting to untangle this hard coupling
is becaise of this.