fix: errors documented as returning from WatchUnitTargetAgentVersion
WatchUnitTargetAgentVersion actually returns a UnitNotFound, instead of
what is documented. The model agent domain had an error on which error
was returned in this case which is also fix. Added tests to the domain
for the various NotFound cases with the various Watch*AgentVersion
methods. More tests should be written for the actual create a watcher
code.