feat: use snaps instead of debs
This is an idea to see if it is feasible to use snaps instead of
debs for our base tooling. We require curl for downloading binaries
(agents etc) and tmux to run debug-hooks. There did use to be a lot
more, but over time I've managed to whittle them down over time. With
ubuntu-fan gone, we can now move over to snaps completely.

Thus, if you've already got a LXD image downloaded, a bootstrap for
me now takes roughly 1 minute. I suspect we could even get that down
further.

This works, because local lxd images are now automatically updated
in the background thanks to this PR[1]. As we're not installing any
PPA repositories or calling apt install, we no longer need to refresh
the source list. Thus, on a moderate connection, we can see a massive
speed boost.

 1. https://github.com/juju/juju/pull/17403