test: use mock for caas storage provisioner tests
The caas storage provisioner are full end to end integration tests
from the apiserver down. There is a roadblock atm were the state
requires domain services and the domain services require state.
This fine because of lazy evaluation at runtime, but is not fine
when for tests.

Moving to mocks removes this problem.