fix(charm relation): fix charm relation DDL
Relations on charms have unique names, even if they are of different
kinds. However, our uniqueness constraint (i.e. the primary key)
included the kind. Drop the kind.

Uniqueness of relation names has been enforced for a very long time:
https://github.com/juju/charm/blob/6b348d6033da7feecfc7272c6eb752b2e8df2e1e/meta.go#L786-L788

Also, as a flyby, add a uuid primary key to the charm relation table. It
will soon be the target of a foriegn key relation.

This meant replacing the old multi-entry primary key mentioned above
with a unique index. This required some changes to the service code.