refactor: fix model defaults state
This commit does several tasks in the model defaults state. The first is
that it adds a new SetCloudDefaults method that will override and set
the a clouds defaults. This is needed to support setting cloud defaults
at bootstrap time.

Previously setting cloud defaults was done in the cloud domain state but
it makes sense for this to be moved to model defaults now like the rest
of of cloud defaults.

While here I have also addresses a number of critial floors in the logic
of this state and fixed several bugs. This is to better align the design
with the Juju teams current values around migrating to DQlite.