feat: implement charm retrieval by hash on domain
Before this patch we had the logic for retrieving a charm by its sha256
hash implemented both in the apiserver and state layers. This is now
moved to the charm (sub-)domain.

In the new implementation, the service layer is in charge of finding the
charm's archive path from the DB using its hash, and then get a reader
by accessing the object store with that archive path. This means that
the object store is now accessed in the service layer (it was already
available on the service layer) and the migration package (it was made
available in https://github.com/juju/juju/pull/18338.