feat(migrations): wire object store getter for import/export operations
Before this patch, the object store getter was not available for
migration operations (as registered in
domain/modelmigration/import|export.go), it was only available for the
domain services.

Since a following patch will include new functionality on the charm
domain that will require having access to an object store, and it will
be passed as argument to the service constructor, it's better to have it
prepared and already wired and available on the migration operations as
well.