fix: fix ssh into noble controllers
The 'juju ssh' command fails to secure a connection to noble
controllers due to a host key verification error

During cloud-init, we re-generate the host keys on a controller.

However, starting with noble, the sshd service has failed to pick up the
newly generated host keys, to 'ssh-keyscan' returns the old, incorrect
host key. This leads to a verification error for 'juju ssh', which
looks up the expected host key from mongo.

This can be verified by running ssh-keyscan and comparing the output to
the key in mongo for both a jammy and noble controller.

I am not exactly sure why this has broken in noble specifically.
However, between jammy and noble sshd become socket-activated, which may
have lead to this.

The solution I have decided on is to reload the service after the new
keys are generated suring cloud-init. Reloading a service after it's
config is changed is a pretty standard operation.

Since this will be performed during cloud-init, I am not worried about
any active ssh connections being severed.