refactor: remove state charm func
There was a mindless translation function that has existed for the
last 8 years[1], which typed cast from a generalised Charm interface
to a *state.Charm. It could do this because it knew that there was
only one shim. Now that we have two, this becomes impossible to
fulfill. The type conversion[2] caused a panic when it was the
other type of shim. This was 80% juju!

Anyway, turns out we can brute force our way through the type system
and force state to work on interfaces and not concrete types.

 1. https://github.com/juju/juju/blame/5e972995681ae74a95bef063a31e4e253e6157c7/apiserver/facades/client/application/application.go#L97-L101
 2. https://github.com/juju/juju/blob/5e972995681ae74a95bef063a31e4e253e6157c7/apiserver/facades/client/application/backend.go#L272-L278