fix(simpleconnector): fix connect() method of simple connector to handle DialOptions
The dial options within the simple connector were not being applied as we were copying the
defaultDialOpts to opts, running the configuration function over a pointer to the copy but passing
the defaultDialOpts. We now just pass the opts. As a driveby, we now allow any api.OpewnFunc to be
passed to Connect.