fix: return resource_oci_image_metadata_store to status quo
For the time being, resource_oci_image_metadata_store will mimic the old
dockerMetadataDoc data with username and password as plain text. A new epic
to handle sensative data, like passwords, which must be persisted is
being created. This password is not the sole place where this is a
concern.