fix: provider defaults on hydration
This commit introduces a fix for provider based model config defaults.
We had introduced a bug into 4.0 when moving model defaults from Mongo
to Dqlite in that when we were hydrating model defaults for a cloud
region or cloud to their actual schema value we were injecting the hard
coded defaults from the provider into the process.

This is not the desired behaviour as the logic only wants to coerce the
values that have been set by the user and not introduce new values.