fix(azure): fix azure provider panic
We previosuly noticed a panic in the azure provider. The initial fix for
this, however, was incorrect.

We believed that recordAuthHeaderPolicy response included a nil header,
which paniced when .Get was called.

However, .Get is safe to call on nil headers. The problem was that is
you don't have network connctivity resp itself was nil, causing a panic
when dereferencing to lookup Headers.

Fix our condition to reflect this