feat: implement file resource store
Implement the file resource store for the resource store factory.

The resource store factory will be used to get the relevent resource
store for the resource type. The interface matchs that of the object
store, though different resource types can be backed by different
stores.

this will be needed for contain image resources.
The resource store put method returns the UUID from the object store
metadata table, this can be used for referential integrirty in the
resource table.