feat: return uuid from objectstore put
Returns a new objectstore UUID from the put service. This requires
wiring the UUID back through the numerous objectstore implementations.

I thought this would be a lot more simpler, but the service for the
objectstore is nestled in the workers. As all the workers are
sequenced, we need to ensure that we return the right information as
well.