feat: remove charm url view
To accommodate annotations in charms, we introduced a charm_url
view, to allow backwards compatibility. We need to stop using that
as it can confuse people when trying to wire up a charm in various
parts of the code.

This solution although not pretty is more correct.

The non-pretty part is the fact we're using ParseURL in the core
package. Once we move that package to the domain, this becomes a lot
more palatable.