fix: ensure we do not leak errors
In trying to optimise the IsError function, we drop functionality,
which meant that we leaked certain types of database errors. The
solution is to just have the correct code, even if it looks
duplicated.

Luckily, there aren't that many calls sites that are using the
database error checking outside of the transaction function (I
did a prior pass sometime ago) and there was a few stragglers.