feat: rework annocations for charms
The annotations for charms required too much magic, and required
the understanding of how a charm might give you a map for querying.
This was too esoteric.

The solution **for now** is to just switch based on the charm type
and push that into the service. This should have been the original
way. It was too clever for it's own good.

The switch statement should be in the service, but this works for
now.