refactor: ports schema to hold RI with charm_relation
At the time, remove the unit_endpoint table. This table wasn't providing
us with anything but pain. Instead, attach the foriegn keys (unit and
charm_relation) directly to the port range

This then required a fairly substantial refactor of the state layer.

Introducing RI from ports to relation resulted in an additional
complexity with the wildcard endpoint, as this does not have a
corresponding charm relation. I have dealt with this by using the null
value. So a null value for relation_uuid represents the wildcard
endpoint. This turned out to be a fairly natural solution.

This refactor also meant that the watcher now emits unit uuids, rather
than unit_endpoint uuids. This meant the watcher code needed a
refactor/simplification as well.