test: watch applications with pending charms
The trigger watches all applications, not just the ones with
pending charms. We need to ensure that the order the changes went
in are the same that come out (there is a dependency chain here!).

The following just tests that with applications that are no longer
pending charms and out of order sequence.