feat(charm): add source selector when retrieving charmID
Before this patch the charm source (local or charmhub) was not part of
the query when retrieving a charm ID by name (plus revision). Now we
also have to provide the charm source.

This work is needed because in a following patch we are going to be
retrieving charms by URL (as URLs are the natural key on the legacy
state) and the charm source (URL schema) was the last missing part
preventing us doing so.