chore: remove origin header
The Origin header is primarily used by web browsers to indicate the origin (scheme, host, and port) of the request. It's a part of the browser's Cross-Origin Resource Sharing (CORS) mechanism to help servers determine the source of a request.

Servers often use this to enforce security policies, like preventing cross-origin attacks.