fix(ec2): remove auto assigned public IP when constraint is false
Before this patch, when using the `allocate-public-ip` constraint set to
false on AWS we still got a public IP(v4) on the instance.

This happens because we accept the default NIC that's created when we
start a new instance, which is taken from the current VPC and subnets,
and by default the network interfaces are associated a public IP.

Since specifying not to assign a public IP to the requested network
interface would have required deep changes in how the interfaces are
created (i.e. instead of accepting the default one, we would be creating
one based on the subnets per AZ, so this incurs into discovering the
subnets first), the solution was to simply remove the public IP of any
created NIC when the constraint is set to false.

The AWS dependencies were updated to make use of the required fields
that allows us to remove the public IP of a NIC
(`AssociatePublicIpAddress`).