refactor: application domain
The application domain was getting really hard to construct, now
with the resources. The solution was to remove all namespaces and
just have a State and a Service, which can be constructed. This
has worked well for all the other domains. We can come back to this
and split this up if we find pieces that can be broken off.