feat: model state implementation
This commit introduces the model state needed for communicating the
model's current state that will be used when calculating a model's
status.

This commit is almost feature complete with the exception of migration
status. Model migration has not been fully pushed into the DQlite domain
yet and so we can't extract what this will be.