refactor(migration): to use domain to migrate charm blobs
Previosuly, we would make an API request to ourselves to retrieve a
charm blob to migrate to a new controller. This is expensive both in
terms of resources, and architectural complexity.

Instead, use the application domain to pull the charm directly from the
object store.

Wire up a CharmService into migrationmaster and refactor
internal/migration to get the archive from a provided service instead.