chore(apiserver): remove unused handler
Now we have cutover migration, the old charms api endpoint is no longer
used. Remove this endpoint and it's handler

As a result, we drop some state calls. Particularly a call to .Charm

As a flyby, remove the object store getter attributes from the object
handler. Instead, get the object store from a helper function directly
from the context. The same way we get state