refactor(actions): deploy local charm in migrate action
We recently broke migrations for charmhub charms. The reason for this is
we do not yet fill in all data for charmhub charms into DQlite that we
do Mongo.

However, we do for local charms. To ensure the action can pass, whilst
retaining signal, download a local charm and deploy that instead.