refactor: change retrived_by to added_by
The field captures the id of the entity that added the resource to
state. If it is a charmhub resource, this will be the app or the unit,
but if it is a user added resource this will be the user name.
retrieved_by didn't quite make sense in the context of the user,
added_by does more so.