fix: model machine info checks controller model machines
Before this patch, any calls to MachineModelInfo() (located in
apiserver/common) would have actually be returning info from the
controller model machines.
This was clearly reproducible if you had a model with 2 machines (and a
no-ha controller), then a call to `juju show-model` would return with an
error stating that machine "1" wasn't found. Another way of reproducing
was to migrate that model.

The fix is to inject the correct MachineService on the modelmanager
facade.