feat: hash 256 and 512-384
The object store currently stores all hashes as 512_384, which is an
issue when trying to search for files based on the hash. The API
objects expect to search via 256. This is because the http handlers
are moving towards s3 compatability. Unfortunately, the backend of
juju uses 512_348. There isn't anything we can do about that right
now, but to make our lives easier for now and in the future, we
can store both. This means if we did want to drop either, it should
be doable now.