chore: remove call to legacy Charm() on bootstrap
In order to start cleaning the legacy state out of Charm() calls, we
need to first remove the one on the bootstrap worker so that Juju is
able to deploy the controller charm without retrieving the charm from
the legacy state.

This is possible because both for local and charmhub charms we have
everything we need (namely URL and Meta), and therefore the only big
change in this patch is to add the CharmURL to the attributes on
AddApplication() and to move the URL() method out from CharmRef to
CharmRefFull, this way we can pass everything during bootstrap.