fix: display_clouds assess clouds test
This test was failing because the json recieved from the `show-clouds`
command was being sorted with JQ. show-clouds was returning a list with
one element and the JQ command was written to work on the first element.
The fix was to first get the first element of the list and then parse.

I am not sure what changed to cause this test to fail, it passed in 3.6
and none of the commits inbetween appear to affect this. Prehaps the
version of JQ was updated somewhere?