feat: implement an application table notify watcher
This PR implements a notify watcher for the application table. This will
be used in the future for the uniter to watch the application table. The
charm_modified_version will be incremented when new resources are added
and the uniter needs to see this in DQLite rather than in state.