fix: allow setting provisioning info for dying machine
We have a potential race where a machine can transition to "dying" while
it is being provisioned. When this happens the provisioner can not
record provisioning info against the machine. If this occurs during
model destruction, we are effectively blocked, because the provisioner
can't come back around to deprovisioning the machine and allowing its
life cycle to advance.

This softens the SetProvisioned checks to ensure a machine is not dead,
instead of not dying or dead. This should give the provisioner the
change to record the instance data, then come around immediately to
deprovision.