feat: remove resources from mongo state functionality
To avoid double writing into a database, remove resources functionality from
mongo state. This change will cause machines charms to not be able to use
file resources. Nor will kubnernetes charms complete deployment.
Lastly juju bootstrap of kubernes may fail as well.

This functionality is being replaced by the resource domain. Work to
wire up that domain and restore functionality is priority.

The following functionality has been temporarily disabled
* resource opener
* the resource client facade
* the resourceshookcontext facade

These will be enabled once they are using the resource domain.

In other facades, the resource methods now return a NotImplemented
error.

In many places new interfaces were created to facilitate the removal of
state.Resources and enable mock tests in the future. A noopShim was used
to return not implemented or nil depending on the code paths.

IncrementCharmModifiedVersion type is gone, facade methods now take a
bool for easy of state removal. It will not be needed in the future due
to changing resource methods in the domain.

Removed the resourcesUploadSuite unit tests in the apiserver. They will
be rewritten for the resource domain and mocks, thus not worth fixing here.