feat: wire up the uniter with the domain application watcher
Before this change, the uniter API embedded the AgentEntityWatcher. This
provided the Watch method on the api and would look up the entity in
state and return a notify watcher.

The attach-resource command increments the CharmModifiedVersion in state
and therefor triggers the watcher.

With resources being moved to DQLite, and the charm_modified_version
integer now being in the application table in domain, this watcher needs
to be watching this table instead.

Not all application functionality has yet been moved over to DQLite.
This could pose an issue, as the uniter checks the CharmModifiedVersion
and the CharmURL when the watcher is triggered. The charm URL should
work as expected as this is currently written in DQLite.

The CharmModifiedVersion is incremented in state in two places. I found
this via the incCharmModifiedVersionOps method and couldn't find any
other places it was set. They are:

    When a resource is set - this will be take care of in the upcoming
    resources work.
    When a charm is changed with SetCharm (there are two callsites in this
    method to incCharmModifiedVersionOps) - The only call remaining for
    this method is doulbe written to DQLite.

Watching the application table in DQLite instead of these should be safe
for the reasons above.

The change made is to remove the AgentEntityWatcher from the UniterAPI
and implement a Watch and WatchApplication on the UniterAPI directly.
The Watch method is kept only so that the uniter API can work with 3.x
clients.

The ApplicationWatcher uses the watcher from the application service,
and the Watch method checks if the tag is an application, if it is,
the Watcher used is from the application service. If it is a unit
watcher, it still watches mongo state.

The uniter call to CharmModifiedVersion is also updated to return the
charm modifed version from the DQLite application table rather than
state.