feat: remove the uniter logic
The fix to the uniter was insane. If the charm wasn't available it
would block for 60 seconds (worst case for each retry, not including
the method in there, so it could be worse!). If there are multiple
charm URLs passed in (bulk API), then it a multiple of the number
of charm URLs.

The fix is simple, although very complex to understand. The uniter
has a downloader (called bundles, which is a legacy concept for
a charm archive)[1], and if it failed it would cause the uniter
to bounce, **except** that the downloader is wrapped in a retrying
bundle reader[2], which correctly backs of on the **caller** side.

Thus, we don't need the retry logic in the facades, just return
the right error code.

Ideally, we wouldn't do this here, with the only other case being
when we download the object. Except that only takes a SHA256 and
if there are multiple charms using the same SHA256, there is no
guarantee that the charms will be all marked as available at the
same time because of the way the async charm downloader worker
works. Because of this race, it makes it hard to guarantee that
the uniter will become available as soon as possible. Essentially,
lock stepping units to the last available charm in the group.
This will become evident in large deployments, esp. with
subordinates.

 1. https://github.com/juju/juju/blob/main/internal/worker/uniter/charm/bundles.go#L73
 2. https://github.com/juju/juju/blob/ae32a0a2f2ce9c29f1eaa8ed40110dbf3322fd04/internal/worker/uniter/charm/manifest_deployer.go#L244-L279