feat: use struct for resource storage ID
The resource storeage ID can either be an object store UUID or a
container metadata image store ID.

Before this change, the UUID was a string, this meant that the
information as to which of the two IDs it was was lost. This change adds
that information back in and makes users of the type be explicit about
which one they expect to get and which one they are saving. Ideally we
would have an enum of the two types, but this is about as close as Go
allows us to get.