chore: handle charmhub identifier
The charmhub identifier should be always set, so that we can ensure
that we have exactly the right information at the point of deployment.
Except that's not currently the case and the fallout from trying to
set it for all locations is problematic.

The solution is to hide it for methods that don't require it and
pass it for when we do. This is passive programming at best. We
should revisit this mess at a later date.