feat: implement charm sequence
Local charms need a sequence table, the sequence table allows for
selecting the next revision for the charm. As charm rows are unique
for each revision; an update to a charm becomes a new row. Then a
new revision is saught. The revision can only be monotonically
increasing, for example: add A, add B, remove B, add C, should result
in 1, 2, 3 and not 1, 2, 2.

We can't get RI for the charm_sequence, as it's inserted first before
the charm. Potentially, this could be solved with a temp view and
the FK could be just on the source tables, though this seems overly
complex for, give me a number in a sequence.