feat(relations): add ddl for relations
First step in moving juju relations data from mongo to dqlite.

application_endpoint maps an application's relation to the space to be
used for new units.

A relation_endpoint maps a relation to an application_endpoint and
provides a uuid to find the application's relation settings as only
one application is referenced. Depending on the relation type, more than
one row per relation is possible. Peer relations will have one row;
relations between two applications will have two rows.

A relation holds life and the relation ID. The relation ID is used by
charms to find the specific relation. A unit may have more than one
relation per endpoint.

A relation_unit relates a unit to a relations and documents whether a
unit is in scope. I.e. has the unit started using the relation?

relation_unit_settings are setting stored by the unit for a specific
relation.

relation_status links a relation to it's status, and the potential
reason for why a relation may have been suspended.

relation_sequence holds the integer for a relation ID. These numbers are
unique to a model and only increment. A digit will not be reused.

There is no scope type or table as it can be inferred from the
application endpoint.