refactor: charm upload logic
Refactor to simplify the apiserver charm upload endpoint.
Particularly the function RepackageAndUploadCharm.

This endpoint does way too much, and does so inefficiently. We
unnecessarily write the charm archive to disk, and read/parse it
unnecessarily.

Worst of all, we even unzip and zip the archive unnecessarily under
most circumstances.

Rework this endpoint to remove redunandency.

The method RepackageAndUploadCharm is also called in the bootstrap
dpeloyer, to upload the controller charm. This is massively overkill.
Drop the dependency and re-write.

At the moment, we do not have methods on the application service to
process local charm blobs. Those are coming soon. So for the meanwhile,
keep with the deprecated state methods.