refactor: remove usages of mongo model status
Removes usages of status for a model from mongo state out of the model
manager api. This work is done as part of the move to DQlite. There are
several failing mock expect tests as part of this. We are skipping them
for now as most of this testing will move down to the services layer
over coming months.