refactor: remove get essential metadata
Both resolve for channel and resolve for deploy return almost
identical information. The former just blanks out the charmhub ID
and the hash. This vestigial from the old state package and all
the information was processed in exactly the same way, just will
different return types. This will hopefully reduce the size of the
repo type.