refactor: remove charm set revision
The revision is set on a local charm to seed the charm archive and
then is read from archive to populate the database. At no point in
the juju code base is the revision then read from the archive
instead of the database.

It's possible for a charmer to read the revision from disk, except
it's not consistent from local charms to charmhub charms. If they
really did want to get the revision, it would be better to set the
revision in a hook envvar and pass that through. It would be more
"correct".