refactor(charm): remove SetDiskRevision
We have recently decided to make charms immutable, especially their
revisions. We made "SetRevision" private.

However, we still exposed "SetDiskRevision", on CharmDir which did
the same thing (as well as writing the new revision to the dir revision
file).

In some places, this was just used as a proxy to set the revision as
usual.

Now that we don't repackage charms on upload, instead relying on
therevision in metadata, there are very few places we reasonable need to
set the charm revision on disk. Even tests.

These places are limited to how we deal with local charm uploads,
because we fall back the the hard-coded revision in the archive if we
don't supply a revision in the charm url.

Refactor tests to remove tautological calls to SetDiskRevision. In the
client/api tests where we do need this, explicitly write the revision to
disk ourselves before the charm is parsed.

As a flyby, since it's closely connected, removed unsupported code paths
in the localcharmclient, where we process charm dirs for local upload.
We no longer support deploying local charms from directory.