test: building dqlite for s390x and ppc64el
Both s390x and ppc64el was removed to allow us to progress the
update of dqlite, and allow time to figure out how to fix the other
architectures.

The issue was that the templating in jenkins has changed and we no
longer encode the arch correctly. This was fixed in jenkins directly
and allows us to build the other two arches.

See: https://jenkins.juju.canonical.com/job/build-dqlite/68/