refactor: bring the environschema package into juju/juju/internal
The environschema package contains the attriutes for the model config
schema. There is no reason for it to live outside Juju. This change
brings it in to the internal package.

This will make autogenerating docs for the schemas that depend on this
package easier.

Note that the external package cannot be entierly expunged from juju as
it is used in the function types of the external idmclient package used
for macaroon authentication:
https://github.com/juju/idmclient/blob/master/ussologin/ussologin.go.
Until this is brought into main we will still need the external
enviromentschema package in the go.mod.

Note that I decided to do this all in one go since the number of files
brought in is relativly small, and so is the number of imports changed.
For a larger package this would be better done in two seperate PRs (lift
and shift).