refactor(annotations): refactor queries into the same file as state
Refactor the annotations domain to include queries in the same file as
the state methods. Also remove some of the conditional logic where it
was no longer necessery.

This change was motivated by an update to SQLair that checks if an
argument to prepare is used. In this domain, different queries were
returns from functions then used with a single prepare. This refator
prepares all the queries next to their declarations. This means the
types can be easily matched to the query.