fix(apiserver): pass the right machine service to modelstatus
This change is necessary because destroying model is broken: if a model has more
machine than the controller model, it won't be destroyed due to an error.
This error is that a call to model status try to get the UUID of a machine,
and get rejected because it use to call the controller's machine service (instead
of the model-to-be-destroyed machine service), leading to error if there is more
machine in model than in controller.

- Before this commit, the `MachineService` was passed directly to ModelStatusAPIs,
  which was wrong since ModelStatus have endpoint addressing multiple model, and
  machine service is bounded to a unique model.
- After this commit, machine service access is now retrieved dynamically using a
  model-specific getter function, allowing ModelStatus to correctly handle multiple
  model requests.